Handle selection for RightClick menu
